Move meeting/agenda_item number tests.
Move to test class that tests corresponding agenda item. Implement for proposal
based and ad-hoc item if not present yet.
