+++ opengever/meeting/tests/test_agendaitem.py
@@ -215,36 +215,6 @@ class TestDecideAgendaItem(IntegrationTestCase):
         agenda_item.decide()
 
     @browsing
-    def test_decide_proposal_agenda_item(self, browser):
-        self.login(self.committee_responsible, browser)
-        agenda_item = self.schedule_proposal(
-            self.meeting, self.submitted_proposal)
-
-        browser.open(self.agenda_item_url(agenda_item, 'decide'),
-                     data={'_authenticator': createToken()})
-
-        self.assertEquals('decided', agenda_item.workflow_state)
-        self.assertEquals([{u'message': u'Agenda Item decided and excerpt '
-                                        u'generated.',
-                            u'messageClass': u'info',
-                            u'messageTitle': u'Information'}],
-                          browser.json.get('messages'))
-
-    @browsing
-    def test_decide_ad_hoc_agenda_item(self, browser):
-        self.login(self.committee_responsible, browser)
-        agenda_item = self.schedule_ad_hoc(self.meeting, 'Gugus')
-
-        browser.open(self.agenda_item_url(agenda_item, 'decide'),
-                     data={'_authenticator': createToken()})
-
-        self.assertEquals('decided', agenda_item.workflow_state)
-        self.assertEquals([{u'message': u'Agenda Item decided.',
-                            u'messageClass': u'info',
-                            u'messageTitle': u'Information'}],
-                          browser.json.get('messages'))
-
-    @browsing
     def test_raises_not_found_for_invalid_agenda_item_id(self, browser):
         self.login(self.committee_responsible, browser)
