+++ opengever/meeting/tests/test_agendaitem_adhoc.py
@@ -10,6 +10,31 @@ class TestAdHocAgendaItem(IntegrationTestCase):
     features = ('meeting',)
 
     @browsing
+    def test_decide_ad_hoc_agenda_item(self, browser):
+        self.login(self.committee_responsible, browser)
+        agenda_item = self.schedule_ad_hoc(self.meeting, 'Gugus')
+
+        browser.open(self.agenda_item_url(agenda_item, 'decide'),
+                     data={'_authenticator': createToken()})
+
+        self.assertEquals('decided', agenda_item.workflow_state)
+        self.assertEquals([{u'message': u'Agenda Item decided.',
+                            u'messageClass': u'info',
+                            u'messageTitle': u'Information'}],
+                          browser.json.get('messages'))
+
+    @browsing
+    def test_decide_ad_hoc_agenda_item_generates_numbers(self, browser):
+        self.login(self.committee_responsible, browser)
+        agenda_item = self.schedule_ad_hoc(self.meeting, 'Gugus')
+
+        browser.open(self.agenda_item_url(agenda_item, 'decide'),
+                     data={'_authenticator': createToken()})
+
+        self.assertEqual(2, agenda_item.decision_number)
+        self.assertEqual(2, self.meeting.model.meeting_number)
+
+    @browsing
     def test_ad_hoc_document_is_created_from_template(self, browser):
         self.login(self.committee_responsible, browser)
